LOT of change in Black Temple + fixed supersonic creatures when speed == 0.0f  + no reagent cost in duel area + added CREATURE_FLAG_EXTRA_DUEL_WIELD flag + some missing support for multi language gossip menu

+++ src/scripts/outland/coilfang_resevoir/underbog/boss_hungarfen.cpp
@@ -134,7 +134,7 @@ struct mob_underbog_mushroomAI : public ScriptedAI
         }else Shrink_Timer -= diff;
     }
 
-    uint32 message(uint32 id, uint32 data)
+    uint64 message(uint32 id, uint64 data)
     { 
         Stop = (bool)data;  
         return 0;
